pyproject.toml
[tool.poetry]
name = "skypro-dz31"
version = "0.1.0"
description = ""
authors = ["flood <teranys1986@gmail.com>"]
readme = "README.md"
packages = [{include = "skypro_dz31"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"
pytest-django = "^4.5.2"
psycopg = "^3.1.9"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
